requirements.txt
discord.py==1.5.1
python-dotenv==0.14.0
beautifulsoup4==4.9.3
requests==2.24.0
